requirements.txt
bleach==5.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
feedparser==6.0.2
idna==3.3
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.5.0
python-dotenv==0.20.0
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.9
validators==0.19.0
webencodings==0.5.1
